pyproject.toml
[tool.poetry]
name = "freelance_website"
version = "0.1.0"
description = ""
authors = ["Anton Krytskyi <ant.krytskyi@gmail.coom>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.4"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
sqlmodel = "^0.0.22"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
alembic-postgresql-enum = "^1.5.0"
greenlet = "^3.1.1"
sqlalchemy-utils = "^0.41.2"
bcrypt = "^4.2.1"

[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
